Reshuffled the source code to make it organized better
