+++ runtime/kubernetes/client/types.go
@@ -1,0 +1,117 @@
+package client
+
+// Metadata defines api object metadata
+type Metadata struct {
+	Name        string            `json:"name,omitempty"`
+	Namespace   string            `json:"namespace,omitempty"`
+	Version     string            `json:"version,omitempty"`
+	Labels      map[string]string `json:"labels,omitempty"`
+	Annotations map[string]string `json:"annotations,omitempty"`
+}
+
+// ServicePort configures service ports
+type ServicePort struct {
+	Name     string `json:"name,omitempty"`
+	Port     int    `json:"port"`
+	Protocol string `json:"protocol,omitempty"`
+}
+
+// ServiceSpec provides service configuration
+type ServiceSpec struct {
+	Type     string            `json:"type,omitempty"`
+	Selector map[string]string `json:"selector,omitempty"`
+	Ports    []ServicePort     `json:"ports,omitempty"`
+}
+
+type LoadBalancerIngress struct {
+	IP       string `json:"ip,omitempty"`
+	Hostname string `json:"hostname,omitempty"`
+}
+
+type LoadBalancerStatus struct {
+	Ingress []LoadBalancerIngress `json:"ingress,omitempty"`
+}
+
+// ServiceStatus
+type ServiceStatus struct {
+	LoadBalancer LoadBalancerStatus `json:"loadBalancer,omitempty"`
+}
+
+// Service is kubernetes service
+type Service struct {
+	Metadata *Metadata      `json:"metadata"`
+	Spec     *ServiceSpec   `json:"spec,omitempty"`
+	Status   *ServiceStatus `json:"status,omitempty"`
+}
+
+// ServiceList
+type ServiceList struct {
+	Items []Service `json:"items"`
+}
+
+// ContainerPort
+type ContainerPort struct {
+	Name          string `json:"name,omitempty"`
+	HostPort      int    `json:"hostPort,omitempty"`
+	ContainerPort int    `json:"containerPort"`
+	Protocol      string `json:"protocol,omitempty"`
+}
+
+// EnvVar is environment variable
+type EnvVar struct {
+	Name  string `json:"name"`
+	Value string `json:"value,omitempty"`
+}
+
+// Container defined container runtime values
+type Container struct {
+	Name  string          `json:"name"`
+	Image string          `json:"image"`
+	Env   []EnvVar        `json:"env,omitempty"`
+	Ports []ContainerPort `json:"ports,omitempty"`
+}
+
+// PodSpec
+type PodSpec struct {
+	Containers []Container `json:"containers"`
+}
+
+// Template is micro deployment template
+type Template struct {
+	Metadata *Metadata `json:"metadata,omitempty"`
+	PodSpec  *PodSpec  `json:"spec,omitempty"`
+}
+
+// LabelSelector is a label query over a set of resources
+// NOTE: we do not support MatchExpressions at the moment
+type LabelSelector struct {
+	MatchLabels map[string]string `json:"matchLabels,omitempty"`
+}
+
+// DeploymentSpec defines micro deployment spec
+type DeploymentSpec struct {
+	Replicas int            `json:"replicas,omitempty"`
+	Selector *LabelSelector `json:"selector"`
+	Template *Template      `json:"template,omitempty"`
+}
+
+// DeploymentStatus is returned when querying deployment
+type DeploymentStatus struct {
+	Replicas            int `json:"replicas,omitempty"`
+	UpdatedReplicas     int `json:"updatedReplicas,omitempty"`
+	ReadyReplicas       int `json:"readyReplicas,omitempty"`
+	AvailableReplicas   int `json:"availableReplicas,omitempty"`
+	UnavailableReplicas int `json:"unavailableReplicas,omitempty"`
+}
+
+// Deployment is Kubernetes deployment
+type Deployment struct {
+	Metadata *Metadata         `json:"metadata"`
+	Spec     *DeploymentSpec   `json:"spec,omitempty"`
+	Status   *DeploymentStatus `json:"status,omitempty"`
+}
+
+// DeploymentList
+type DeploymentList struct {
+	Items []Deployment `json:"items"`
+}
